+++ ask-sdk-model/CHANGELOG.rst
@@ -291,3 +291,12 @@ This release contains the following changes :
 This release contains the following changes : 
 
 - Updated `rules for recurrence creation <https://developer.amazon.com/en-US/docs/alexa/smapi/alexa-reminders-api-reference.html#in-session-and-out-of-session-behavior-for-alexa-reminders-api>`__ in reminders. 
+
+
+1.22.0
+~~~~~~
+
+This release contains the following changes : 
+
+- APIs related to `timer management <https://developer.amazon.com/en-US/docs/alexa/smapi/alexa-timers-api-reference.html>`__.
+